server/requirements.txt
flask
flask-sqlalchemy
flask-bcrypt
python-dotenv
flask-cors
flask-session
redis 
